+++ EssentialFeedTests/Feed API/URLSessionHTTPClientTests.swift
@@ -124,16 +124,8 @@ final class URLSessionHTTPClientTests: XCTestCase {
         return receivedResult!
     }
     
-    private func anyURL() -> URL {
-        return URL(string: "http://any-url.com")!
-    }
-    
     private func anyData() -> Data {
         Data("any data".utf8)
-    }
-    
-    private func anyNSError() -> NSError {
-        NSError(domain: "any error", code: 1)
     }
     
     private func nonHTTPURLResponse() -> URLResponse {
