+++ EssentialFeedTests/Feed Cache/Helpers/FeedCacheTestHelpers.swift
@@ -1,0 +1,29 @@
+//  Created by Wiiliam Peregoy on 7/25/23
+
+import Foundation
+import EssentialFeed
+
+func uniqueImage() -> FeedImage {
+    FeedImage(
+        id: UUID(),
+        description: "any",
+        location: "any",
+        url: anyURL()
+    )
+}
+
+func uniqueImageFeed() -> (models: [FeedImage], local: [LocalFeedImage]) {
+    let models = [uniqueImage(), uniqueImage()]
+    let local = models.map { LocalFeedImage(id: $0.id, description: $0.description, location: $0.location, url: $0.url) }
+    return (models, local)
+}
+
+extension Date {
+    func adding(days: Int) -> Date {
+        return Calendar(identifier: .gregorian).date(byAdding: .init(day: days), to: self)!
+    }
+    
+    func adding(seconds: TimeInterval) -> Date {
+        return self + seconds
+    }
+}
