+++ EssentialFeedTests/Helpers/SharedTestHelpers.swift
@@ -1,0 +1,12 @@
+//  Created by Wiiliam Peregoy on 7/25/23
+
+import Foundation
+
+func anyNSError() -> NSError {
+    NSError(domain: "any error", code: 1)
+}
+
+func anyURL() -> URL {
+    return URL(string: "http://any-url.com")!
+}
+
